fix: resolve Rust 1.95 compatibility issues
- branch_protection.rs: fix Arc::clone for dyn CiProvider
- file_append.rs: fix into_tree() error type for and_then
- mirror.rs: add explicit lifetime to push_to_target
- performance.rs: fix tcp_keepalive Option<Duration> type
- rate_limiter.rs: fix window_secs u64->u32 conversion
- security.rs: fix Option username/password is_empty checks
- tls.rs: fix rustls_pemfile certs lifetime issues

+++ crates/opengit-core/src/rate_limiter.rs
@@ -509,7 +509,7 @@ impl RateLimiter {
                 allowed: true,
                 remaining: self.config.ip.read_limit as u32,
                 limit: self.config.ip.read_limit as u32,
-                reset_in: self.config.ip.window_secs,
+                reset_in: self.config.ip.window_secs as u32,
             }
         }
     }
@@ -531,7 +531,7 @@ impl RateLimiter {
                 allowed: true,
                 remaining: self.config.identity.read_limit as u32,
                 limit: self.config.identity.read_limit as u32,
-                reset_in: self.config.identity.window_secs,
+                reset_in: self.config.identity.window_secs as u32,
             }
         }
     }
